Rename `ext` module to `gltf_ext` and move it to `loader` module in `bevy_gltf`
