--- crates/bevy_gltf/src/ext/mod.rs
@@ -1,9 +1,0 @@
-//! Extension traits for [`gltf`] types
-
-pub mod document;
-pub mod gltf;
-pub mod json;
-pub mod material;
-pub mod mesh;
-pub mod node;
-pub mod texture;
