+++ crates/bevy_gltf/src/lib.rs
@@ -91,7 +91,6 @@
 //! You can use [`GltfAssetLabel`] to ensure you are using the correct label.
 
 mod assets;
-mod ext;
 mod label;
 mod loader;
 mod vertex_attributes;
