+++ crates/bevy_gltf/src/loader/gltf_ext/extras.rs
