+++ crates/bevy_gltf/src/loader/gltf_ext/mesh.rs
@@ -1,7 +1,12 @@
 use bevy_render::mesh::PrimitiveTopology;
-use gltf::mesh::Mode;
+
+use gltf::mesh::{Mesh, Mode, Primitive};
 
 use crate::GltfError;
+
+pub trait MeshExt {
+    fn primitive_name(&self, primitive: &Primitive) -> String;
+}
 
 pub trait ModeExt {
     #[expect(
@@ -9,6 +14,17 @@ pub trait ModeExt {
         reason = "`GltfError` is only barely past the threshold for large errors."
     )]
     fn primitive_topology(self) -> Result<PrimitiveTopology, GltfError>;
+}
+
+impl MeshExt for Mesh<'_> {
+    fn primitive_name(&self, primitive: &Primitive) -> String {
+        let mesh_name = self.name().unwrap_or("Mesh");
+        if self.primitives().len() > 1 {
+            format!("{}.{}", mesh_name, primitive.index())
+        } else {
+            mesh_name.to_string()
+        }
+    }
 }
 
 impl ModeExt for Mode {
